Fixed compilde error cdecl ParticleManager . Fixed point reset between levels.
1. Fixed error: "Severity Code Description Project File Line Suppression State Details Error LNK2001 unresolved external symbol "public: void __cdecl ParticleManage" by changing in the  properties of file "ParticleManager.h" the item type. From the wrong type c++ compiler to c++ header.

+++ C_Plus_Plus_Game/Level.h
@@ -8,7 +8,7 @@ class Level : public GameObject
 {
 	float m_center_x = 3.0f;
 	float m_center_y = 3.0f;
-	float pointsGainedInLevel=0;
+	int pointsGainedInLevel = 0;
 	//? everything included in the level
 	std::vector<GameObject*> m_static_objects;
 	std::vector<std::string> m_fireball_sprites;
